refactor: rename swaps cta
